IMR-83 refactor : add env setting and separate route from app
