+++ back/requirements.txt
@@ -1,3 +1,4 @@
 fastapi[all]
 pydantic==2.0.2
+python-dotenv==1.0.0
 python-multipart==0.0.6
